# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tal Misri Factory Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cash:
    get:
      operationId: listCashEntries
      tags: [cash]
      summary: List all cash entries
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cash entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashEntry"
    post:
      operationId: createCashEntry
      tags: [cash]
      summary: Create a new cash entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCashEntryBody"
      responses:
        "201":
          description: Created cash entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashEntry"
  /cash/{id}:
    get:
      operationId: getCashEntry
      tags: [cash]
      summary: Get a single cash entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cash entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashEntry"
    put:
      operationId: updateCashEntry
      tags: [cash]
      summary: Update a cash entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCashEntryBody"
      responses:
        "200":
          description: Updated cash entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashEntry"
    delete:
      operationId: deleteCashEntry
      tags: [cash]
      summary: Delete a cash entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cash/summary/daily:
    get:
      operationId: getDailyCashSummary
      tags: [cash]
      summary: Get daily cash summary (opening, income, expense, closing)
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyCashSummary"

  # ── PARTIES ──
...
  /batches:
    get:
      operationId: listBatches
      tags: [production]
      summary: List all production batches
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [boiling, curing, ready, harvested]
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [production]
      summary: Start a new production batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Created batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    get:
      operationId: getBatch
      tags: [production]
      summary: Get a single batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    put:
      operationId: updateBatch
      tags: [production]
      summary: Update batch status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatchBody"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}/harvest:
    post:
      operationId: harvestBatch
      tags: [production]
      summary: Record harvest output for a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HarvestBody"
      responses:
        "200":
          description: Harvested batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"

  # ── STOCK ──
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    CashEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        partyId:
          type: integer
          nullable: true
        partyName:
          type: string
          nullable: true
        buyerName:
          type: string
          nullable: true
        saleItems:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, date, amount, type, category, createdAt]
    CreateCashEntryBody:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        partyId:
          type: integer
          nullable: true
        buyerName:
          type: string
          nullable: true
        saleItems:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: [date, amount, type, category]
    DailyCashSummary:
      type: object
      properties:
        date:
          type: string
        openingBalance:
          type: number
        totalIncome:
          type: number
        totalExpense:
          type: number
        closingBalance:
          type: number
      required: [date, openingBalance, totalIncome, totalExpense, closingBalance]
...
    Batch:
      type: object
      properties:
        id:
          type: integer
        batchCode:
          type: string
        startDate:
          type: string
        expectedEndDate:
          type: string
        sugarInputKg:
          type: number
        traysCount:
          type: integer
        workerName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [boiling, curing, ready, harvested]
        daysRemaining:
          type: integer
          nullable: true
        grade1OutputKg:
          type: number
          nullable: true
        grade2OutputKg:
          type: number
          nullable: true
        byproductKg:
          type: number
          nullable: true
        lossKg:
          type: number
          nullable: true
        yieldPercent:
          type: number
          nullable: true
        harvestedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, batchCode, startDate, expectedEndDate, sugarInputKg, traysCount, status, createdAt]
    CreateBatchBody:
      type: object
      properties:
        startDate:
          type: string
        sugarInputKg:
          type: number
        traysCount:
          type: integer
        workerName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [startDate, sugarInputKg, traysCount]
    UpdateBatchBody:
      type: object
      properties:
        status:
          type: string
          enum: [boiling, curing, ready, harvested]
        notes:
          type: string
          nullable: true
      required: [status]
    HarvestBody:
      type: object
      properties:
        grade1OutputKg:
          type: number
        grade2OutputKg:
          type: number
        byproductKg:
          type: number
      required: [grade1OutputKg, grade2OutputKg, byproductKg]
...