# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Tal Misri Factory Business Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
  message: string;
}