# lib/api-zod/src/generated/types/updateBatchBody.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Tal Misri Factory Business Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBatchBodyStatus } from './updateBatchBodyStatus';

export interface UpdateBatchBody {
  status: UpdateBatchBodyStatus;
  notes?: string | null;
}